dice_deploy_django/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
babel==2.2.0              # via flower
backports.ssl-match-hostname==3.5.0.1  # via tornado
billiard==3.3.0.23        # via celery
celery==3.1.23
certifi==2016.2.28        # via tornado
cloudify-rest-client==3.2.1
coreapi==1.32.0           # via django-rest-swagger, openapi-codec
django-celery==3.1.17
django-rest-swagger==2.0.1
django==1.9.4
djangorestframework==3.4.0
enum34==1.1.2
flower==0.9.0
funcsigs==1.0.2           # via mock
futures==3.0.5            # via flower
gunicorn==19.4.5
itypes==1.1.0             # via coreapi
jsonfield==1.0.3
kombu==3.0.34             # via celery
markdown==2.6.6
mock==2.0.0
openapi-codec==1.0.0      # via django-rest-swagger
pbr==1.10.0               # via mock
pytz==2016.2              # via babel, celery, flower
PyYAML==3.11
requests==2.7.0           # via cloudify-rest-client, coreapi
simplejson==3.8.2         # via django-rest-swagger, uritemplate
six==1.10.0               # via mock
testfixtures==4.10.0
tornado==4.2.0            # via flower
uritemplate==0.6          # via coreapi
